# backend-flask/openapi-3.0.yml
openapi: '3.0.2'
info:
  title: Blinkr Backend API
  version: '1.0'

servers:
  - url: https://api.server.test/v1
paths:
  /api/teams:
    get:
      description: 'Return all teams, or only the ones the signed-in user follows'
      tags:
        - teams
      parameters:
        - name: followed
          in: query
          description: Set to "true" to scope to the signed-in user's followed teams
          required: false
          schema:
            type: string
            example: "true"
      responses:
        '200':
          description: Returns an array of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
  /api/teams/{team_uuid}:
    get:
      description: 'Return a single team and its matches'
      tags:
        - teams
      parameters:
        - name: team_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns a team with a nested matches array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          description: Team not found
  /api/teams/{team_uuid}/follow:
    post:
      description: 'Follow a team as the signed-in user'
      tags:
        - teams
      parameters:
        - name: team_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Follow confirmation
        '401':
          description: Not authenticated
    delete:
      description: 'Unfollow a team as the signed-in user'
      tags:
        - teams
      parameters:
        - name: team_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unfollow confirmation
        '401':
          description: Not authenticated
  /api/matches:
    get:
      description: 'Return all matches, or only the ones the signed-in user follows'
      tags:
        - matches
      parameters:
        - name: followed
          in: query
          description: Set to "true" to scope to the signed-in user's followed matches
          required: false
          schema:
            type: string
            example: "true"
      responses:
        '200':
          description: Returns an array of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
  /api/matches/live:
    get:
      description: 'Return matches currently in progress'
      tags:
        - matches
      responses:
        '200':
          description: Returns an array of live matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
  /api/matches/upcoming:
    get:
      description: 'Return scheduled matches that have not kicked off yet'
      tags:
        - matches
      responses:
        '200':
          description: Returns an array of upcoming matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
  /api/matches/{match_uuid}:
    get:
      description: 'Return a single match'
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns a match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '404':
          description: Match not found
  /api/matches/{match_uuid}/events:
    get:
      description: 'Return the official event timeline for a match (goals, cards, substitutions)'
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns an array of match events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchEvent'
    post:
      description: >
        Record a match event (goal, card, substitution, half/full time).
        Admin/demo tool used to drive match state -- not wired into the
        frontend, and intentionally unauthenticated (there is no admin-role
        system in this project, matching the existing write routes).
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_type:
                  type: string
                  enum: [GOAL, YELLOW_CARD, RED_CARD, SUBSTITUTION, HALF_TIME, FULL_TIME]
                minute:
                  type: integer
                team_uuid:
                  type: string
                player_name:
                  type: string
                detail:
                  type: string
      responses:
        '200':
          description: Returns the created match event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchEvent'
        '422':
          description: Validation error
  /api/matches/{match_uuid}/reactions:
    get:
      description: 'Return active (non-expired) reactions posted during a match'
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns an array of match reactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchReaction'
    post:
      description: 'Post a reaction to a match as the signed-in user'
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        '200':
          description: Returns the created reaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchReaction'
        '401':
          description: Not authenticated
        '422':
          description: Validation error
  /api/matches/{match_uuid}/follow:
    post:
      description: 'Follow a match as the signed-in user'
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Follow confirmation
        '401':
          description: Not authenticated
    delete:
      description: 'Unfollow a match as the signed-in user'
      tags:
        - matches
      parameters:
        - name: match_uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unfollow confirmation
        '401':
          description: Not authenticated

components:
  schemas:
    Team:
      type: object
      properties:
        uuid:
          type: string
          example: 0056a369-4618-43a4-ad88-e7c371bf5582
        name:
          type: string
          example: "FC Barcelona"
        short_name:
          type: string
          example: "Barcelona"
        abbreviation:
          type: string
          example: "BAR"
        country:
          type: string
          example: "Spain"
    Match:
      type: object
      properties:
        uuid:
          type: string
          example: 0056a369-4618-43a4-ad88-e7c371bf5582
        competition:
          type: string
          example: "El Clasico"
        kickoff_time:
          type: string
          example: "2023-02-06T18:11:03+00:00"
        status:
          type: string
          enum: [scheduled, live, half_time, finished]
        home_score:
          type: integer
          example: 1
        away_score:
          type: integer
          example: 1
        elapsed_minute:
          type: integer
          nullable: true
          example: 65
        home_team_uuid:
          type: string
        home_team_name:
          type: string
          example: "Real Madrid CF"
        home_team_abbreviation:
          type: string
          example: "RMA"
        away_team_uuid:
          type: string
        away_team_name:
          type: string
          example: "FC Barcelona"
        away_team_abbreviation:
          type: string
          example: "BAR"
    MatchEvent:
      type: object
      properties:
        uuid:
          type: string
          example: 0056a369-4618-43a4-ad88-e7c371bf5582
        event_type:
          type: string
          enum: [GOAL, YELLOW_CARD, RED_CARD, SUBSTITUTION, HALF_TIME, FULL_TIME]
        minute:
          type: integer
          example: 73
        player_name:
          type: string
          example: "Jude Bellingham"
        detail:
          type: string
          example: "on for Rodrygo"
        team_abbreviation:
          type: string
          example: "RMA"
        created_at:
          type: string
          example: "2023-02-06T18:11:03+00:00"
    MatchReaction:
      type: object
      properties:
        uuid:
          type: string
          example: 0056a369-4618-43a4-ad88-e7c371bf5582
        display_name:
          type: string
          example: "Andrew Brown"
        handle:
          type: string
          example: "andrewbrown"
        message:
          type: string
          example: "What a strike from Bellingham!"
        created_at:
          type: string
          example: "2023-02-06T18:11:03+00:00"
        expires_at:
          type: string
          example: "2023-02-06T21:11:03+00:00"